implement test_block_internals for series
